requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0

supabase>=1.0.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
email-validator>=2.1.0
geopy>=2.3.0
apscheduler>=3.10.4
firebase-admin>=6.0.0
stripe>=7.0.0
redis>=5.0.0
aioredis>=2.0.0
pillow>=10.0.0
python-multipart>=0.0.6
